[FE] refactor: post 컴포넌트 구조 개선
